start 스크립트 변경

+++ web/scripts/createDB.sh
@@ -3,13 +3,17 @@
 sudo apt install sqlite3 
 
 sudo mkdir -p /var/tmp
-sudo touch /var/tmp/prod.sqlite3 
-sudo chmod 755 /var/tmp/prod.sqlite3
-sudo chmod u+w /var/tmp
-sudo chown -R user /var/tmp/prod.sqlite3
-
-
-npm install -g cross-env ts-node
-DB_PATH=/var/tmp/prod.sqlite3 
-NODE_ENV=production 
-npx ts-node '../server/db/createScript.ts'
+if [ -f /var/tmp/prod.sqlite3 ]; then
+  exit 1;
+else
+  sudo touch /var/tmp/prod.sqlite3 
+  sudo chmod 755 /var/tmp/prod.sqlite3
+  sudo chmod u+w /var/tmp
+  sudo chown -R user /var/tmp/prod.sqlite3
+  
+  
+  npm install -g cross-env ts-node
+  DB_PATH=/var/tmp/prod.sqlite3 
+  NODE_ENV=production 
+  npx ts-node '../server/db/createScript.ts'
+fi
